Move some fixtures to conftest.
